After much work the final version is up!
As of now this code COMPLETELY WORKS, but with only text files as of now, but this weekend ill be adding support for both images and excel files.  I have not been able to break the server yet, and it even holds up to junk data/request and random dropping clients. I'll update the image and excel support when I have it fully working