CAN Snooper: Added MiniFitJr-4-peg footprint; regenerated netlist.
The connector we plan to use for power and CAN is a Molex Mini-Fit Jr
with 4 pins (Molex calls them "circuits"). The most compact of those
designs uses "pegs" to secure itself to the board.

I designed a footprint for this connector and assigned it to the two
connector footprints we have on the schematic, then regenerated the
netlist.