MPPT: First stab at recreating the original schematic.
In order to improve upon the previous design which seems to have gone missing, I'm recreating it and reorganizing the components to make the schematic more clear.

The CAN circuitry has been split into a separate breakout board, so it doesn't need to be included in this schematic.

Current progress:
* DC/DC Converter recreated
* Voltage Sensors recreated

Still to do:
* Power Regulation and Isolation
* Current Sensors
* FET driver
* Nucleo connections
* Snubber values and diode direction.