MPPT: Add power lines and regulators
There are three primary power regions on the board: the high power regions around the DC/DC Converter, the car's "low" power regions (+12V), and the low power regions around the Nucleo.

The high power regions include current sensors and FET drivers which require power from the car's low power line rather than the high power lines from the solar array and the battery. We use a 12V-12V DC isolator to provide this safely.

We also isolate power to the Nucleo for safety when attaching a computer to the board. Finally, the CAN lines have an isolated power region, but this is managed by the CAN chip.