MPPT: Add FET driver and its Nucleo connection
The FET driver doesn't provide any isolation, and it needs to be on the High Power plane, so we will need to add isolation on its PWM input to protect the Nucleo.

We should also figure out whether tying the chip's Enable pin to the PWM input is appropriate or not. According to the datasheet, the enable pin is internally pulled up to Vdd by a 100Kohm resistor, so it may draw unexpected current from the PWM line.