Electronics: Add Molex Power and CAN connector footprints; PC Array Power connector footprint
For low-power systems, we're using Molex Micro-fit 3.0 connectors. We have a 3-pin variant for power (+12,+5,GND) and a 4-pin variant for CAN (+5,GND,CANH,CANL).

For the solar array and battery, we're using Phoenix Contact Power Combicon 5 connectors. To ensure the battery and array can't be swapped, we're using a 3-pin connector for the array and a 2-pin connector for the battery.