Electronics: Set default pad clearance on SOIC-6_Pitch0.65mm to 0.08mm
Normally we would leave the default clearance to 0 (i.e. "use global"), but because the pads are so close to each other, the global clearance tends to be far too large. A 0.08mm clearance takes up approximately one third of the distance between the outsides of the pads, so that's a good default.