MPPT: Add labels to connector pins; fix polarity markings on U3,4,6; clarify polarity markings on power diodes
The polarity markings on U3,4,6 were flipped along the long-axis, marking the last pin rather than the first. I've corrected that, importing the changes to the footprints from previous commits.

I changed the polarity markings on the power diodes from +,- to A,C to indicate anode/diode more clearly, especially since the + side may not necessarily be at a higher voltage than the - side.