Repair the current and ADC drivers.
Problem
=======
The current driver does not work.

Solution
========
The main issue appeared to be with the order of the ADC readings.
Apparently, a priority of 1 is read after a priority of 2 when using
DMA, meaning that the sensors we're brought in in the worng order.

Since each sensor was being scaled differently, this means that
one was ending up as twice the proper current and the other was
approximately half. To fix this, the priorities were swapped.

Notes
=====
The ADC driver should now work as expected. Be careful when changing
anything in that file, since changing priorities may alter the
behavior if other things are not also corrected.