Finish EEPROM and I2C
Problem
=======
The EEPROM and I2C libraries were not reading and writing well.
This was partially due to an I2C error that was not accounting
for NAKs on address writes to the EEPROM when the previous
write-cycles was still ongoing. The other part were high-level
EEPROM driver issues that needed to be corrected to ensure
persistence of memory.

Solution
========
The I2C driver was fixed by including a check for NAK when waiting
for the transmission-mode flag to be set after sending a device
address.