Get CAN in a working state.
Problem
=======
CAN was not working correctly. This had to do with an error in the
testcase, a mismatch between the standard peripheral libray and
actual functionality, a timing issue, and mode configuration for
testing.

Solution
========
The following changes we made:
 * Fixed the malloc in the testcase
 * Reconfigured the timing to match with the other systems
 * Added mode configuration in the driver initialization function
 * Created a new `const` in the driver to use with AF configuration.

Testing
=======
These changes have __not__ been tested with the other systems yet.

Note
====
This is another system that we should eventually move to an
interrupt-driven operation, since otherwise we run the risk of
clogging up our processing power with communications polling.