Complete the BPS CAN messages as specified.
Problem
=======
The BPS CAN messages would not be read correctly by the telemetry
system. This was because the BPS CAN messages were being formatted
improperly.

Solution
========
The solution code was implemented from observing the way the telemetry
code parses the data. Importantly, this means that *THE CAN MESSAGES
AS SPECIFIED ON THE SHARED DOCUMENT ARE OUT OF DATE.* These should be
updated at the time this commit is merged into the master to avoid
future confusion.