Refactor the build system with the new Makefile.
Overview
========
This commit extends upon the introduction of the Linux
Makefile to make the build-system much cleaner. This is
done by:

1. Moving the "real" `main()` in `main.c` to be located at
the bottom of the file. In addition, the systemLoop function
was introduced at the top of the file for basic organization.
2. The creation of the `solar-build` script, which negates the
need to have the definition for the test that we want to run
in the actual files. In this way, we can build multiple targets
without needing to modify files.
3. Updating the README.md with instructions on how to get the
new build system working on a bash shell (assuming a Debian/Ubuntu
installation).
4. Fixing up some minor things to get the thing to build right.