Make the party parrot a separate target.
The party parrot now requires a definition
to be in place in order to be included in
the build.

When the definition is in place, the binary
is approximately 47 KiB. when the definition
is not in place (so no party parrot), the
binary is approximately 37 KiB. this is about
a 10KiB saving by removing the parrot.

I also fixed up a weird function defintion
that had an extra parameter.