Fix terminal badness.
Problem
=======
When running the program improperly, the terminal is FUBAR. This
requires a restart.

Solution
========
Add a linked file in the root of the BeVolt folder that can be used
to run the program, add a general case for error handling to avoid
the terminal formatting issue caused by a failure to end ncurses.