Add interrupt management functions and a simulation flag
Problem
=======
Commented code bugs me. It should be deleted or used.

Solution
========
I uncommented the __enable_irq() and __disable_irq() functions
in main.c. This only works in the ARM assembly syntax. In order
to determine whether we are targetting ARM or x86 at compile-time,
and additional flag (SIMULATION) has been added, such that it being
defined implies that we are targetting an x86 (or at the very least,
a non-ARM) architecture.