Fix the Makefile's flash and clean functionalities.
Problem
=======
This has been bugging me. Clean should not spit out
error messages unless something has actually gone wrong.
A file not being there when it needs to be removed is
a non-problem.

This was also missing the ability to flash to the board.

Solution
========
It's pretty simple changes. Take a look yourself.