Fix CAN file access error and improve CAN test case.
This fixes the issue present in CAN.py that was causing intermittent
crashes of the simulator. The test case has also been changed to one
that exposed this issue on the previous version of the simulator.

On the old version of the code, crashes were observed anywhere from
a few seconds after simulator launch to about one minute after launch.
After the change, I have successfully run the simulator for over 2 hours
without a crash occurring.