Change the 'SoC' to 'Charge' and make implementation nicer.
Problem
=======
SoC commonly refers to System on a Chip, but we have been using it
to mean State of Charge for a while. This is somewhat confusing
terminology.

Solution
========
Renamed (almost) everything from SoC to Charge. This includes everything
in the SoC.h and SoC.c files (which have been renamed Charge.h and Charge.c).
This does not include definitions provided in the EEPROM and CAN files.
I excluded the CAN file because I did not want to rename CAN IDs
without input from everyone.

While addressing this problem, I also came across some things in the
(formerly) SoC file that made me a bit nervous, so I went ahead and
made changes that I thought were appropriate.

The biggest of these changes was the elimination of the two variables
tracking the state of charge. For some reason, the app was tracking
charge via a global float and a global int, and only one was ever
accessed outside the file. It now just uses and int.

Additionally, I fixed up the units on the file, which I believe were
inconsistent with the usage with the other modules. Additionally,
I believe that I found a bug having to do with passing updates
into the Charge update function, which has now been fixed.

Also, I added -j4 to the Makefile build targets so we get a speedier build.