Add a preliminary driver for the Ampere minion.
Important
=========
This is __not__ compatible with the simulator in its present state.
This would require adding a SPI3 simulator, which I am not including
in this commit.

Problem
=======
We're replacing the existing analog minion board with a digital one.
This requires new drivers.

Solution
========
This is my best guess at what an interface to the Ampere minion would
look like, without actually having access to the hardware to test. It
uses SPI3 to communicate to the minion, and includes what I think is
necessary initialization.

Notes
=====
This also renames the Current.c/Current.h files and the associated
functions to Amps.c/Amps.h, relating to a decision to use less
confusing terminology.