Refactor SPI BSP code to support multiple busses.
Problem
=======
We want to be able to control multiple SPI busses without
duplicating code each time.

Solution
========
We now have a single SPI BSP source that supports multiple
busses, with minimal code duplication, which now only takes
place in the initialization function (and is rather unavoidable).

Notes
=====
The BSP for the simulator compiles, but is not complete, and so the
simulator cannot be used for this as of yet.