CAN Interrupts and OS Integration (#226)
* Restructure CAN BSP and driver to interrupt/OS.

Problem
=======
The CAN driver and BSP code was previously using polling-like
techniques to send and receive messages. This will not integrate
well with the move to uC/OS.

Solution
========
The CAN drivers are being moved to an interrupt-driven system,
whereby the receiving messages come via interrupt for later collection,
and the transmitted messages utilize interrupts and locks to ensure
thread safety.

There are now two methods for transmitting messages. These are:
* CANbus_Send, which preserves closely its old functionality for backwards compatibility.
* CANbus_BlockAndSend, which puts the calling thread to sleep until the message can be sent.

If the first of the above messages is called, the caller is expected
to check for a proper transmission. If a guarenteed transmission is
required, the second function should be called.

To receive messages, there are likewise two functions: one nonblocking and
one blocking (CANbus_Receive & CANbus_WaitToReceive). Use the first one
only if you do not want to block and do not plan to poll.

Notes
=====
No testing has been done on these yet, and there has been no work on the
simulator portions.

* Restructure CAN BSP and driver to interrupt/OS.

Problem
=======
The CAN driver and BSP code was previously using polling-like
techniques to send and receive messages. This will not integrate
well with the move to uC/OS.

Solution
========
The CAN drivers are being moved to an interrupt-driven system,
whereby the receiving messages come via interrupt for later collection,
and the transmitted messages utilize interrupts and locks to ensure
thread safety.

There are now two methods for transmitting messages. These are:
* CANbus_Send, which preserves closely its old functionality for backwards compatibility.
* CANbus_BlockAndSend, which puts the calling thread to sleep until the message can be sent.

If the first of the above messages is called, the caller is expected
to check for a proper transmission. If a guarenteed transmission is
required, the second function should be called.

To receive messages, there are likewise two functions: one nonblocking and
one blocking (CANbus_Receive & CANbus_WaitToReceive). Use the first one
only if you do not want to block and do not plan to poll.

Notes
=====
No testing has been done on these yet, and there has been no work on the
simulator portions.

* Fix simulator compilation errors. Simulator still not implemented.w

* changed CAN functions so they return ErrorStatus instead of uint8_t or int

* Fixed simulator build error

Co-authored-by: ClarkPoon <clarkap26@gmail.com>