Remove warnings in the code. Add -Werror flag.
There's a lot of warnings here, and we want to remove them. To make
sure that we never have to do a commit like this again, this also
adds the -Werror flag to both Makefiles.

As a note, there were also some logic errors in the code as well.
Fixing the warnings also brought some attention to fixing these.