Warnings (#263)
* Add handling of a fault state. Also fix some simulator compile errors.

* Remove warnings in the code. Add -Werror flag.

There's a lot of warnings here, and we want to remove them. To make
sure that we never have to do a commit like this again, this also
adds the -Werror flag to both Makefiles.

As a note, there were also some logic errors in the code as well.
Fixing the warnings also brought some attention to fixing these.

* Remove warnings in the code. Add -Werror flag.

There's a lot of warnings here, and we want to remove them. To make
sure that we never have to do a commit like this again, this also
adds the -Werror flag to both Makefiles.

As a note, there were also some logic errors in the code as well.
Fixing the warnings also brought some attention to fixing these.

Co-authored-by: ClarkPoon <43055250+ClarkPoon@users.noreply.github.com>