Test as8510 (#384)
* work in progress

* work in progress

* got BAREMETAL working

* work in progress

* I'm pretty sure it's working, but interrupts don't seem to help performance

* Modify the fifo and spi to do interrupts better.

We weren't really taking advantage of them with the older
code

* Get SPI working.

This commit encompasses quite a bit:
    - Modifies the SPI interrupt method to take better advantage
      of the time that interrupts give us.
    - Modifies some test cases to appropriately test SPI and the LTC6811.
    - Fixes a bug in the LTC6811 driver (not initializing a semaphore).

The interrupt have been modified to buffer data being written/read in
software fifos that the interrupts then use to pick data from and stash
data to.

As a note: the SPI read and write methods are not thread safe if two threads
are attempting to access the same SPI port. Different ports should be OK.

When running the LTC6811 test program, the following output is given:
```
Testing LTC6811 functions.
pec: 0
Minion 0:
        Voltages:
        0: 4199mV
        1: 4194mV
        2: 4219mV
        3: 4216mV
        4: 4191mV
        5: 4215mV
        6: 0mV
        7: 0mV
        8: 0mV
        9: 0mV
        10: 0mV
        11: 0mV
```
The hardware is set up such that only the lower 6 cells have a voltage
potential applied to them. Measuring with a multimeter confirms that these
readings seem accurate (although exact accuracy is not being confirmed at this
time). Both the interrupt-driven method and the polling backup read about the
same.

Be aware that the measurements seem to degrade when the bitrate of the SPI bus
is increased, so this commit also includes setting the prescalar to 256 from its
original value of 64.

* work in progress

* work in progress

* work in progress

* fixed SPI3

* work in progress

* work in progress

* added chip select to test program

* fixed NUM_MINIONS in config.h

* added formula to convert from AS8510 ADC readings to milliamps

* completed AS8510 test plan

* fix compiler warning

Co-authored-by: Chase Block <chase@chaseblock.com>